refactor: extract agent-detail helpers, fix pulse circular import
Move form renderers, model data, and small helpers from
agent-detail-v2.ts (536 → 393 lines) into agent-detail-helpers.ts.
Extract PulseTile/TileWrapFn types into pulse-types.ts to break
the circular import between pulse.ts and pulse-renderers.ts.
